+++ lizard_auth_client/testsettings.py
@@ -104,7 +104,7 @@ SSO_SECRET = 'random_generated_secret_key_to_sign_exchanged_messages'
 # URL used to redirect the user to the SSO server
 # Note: needs a trailing slash
 SSO_SERVER_PUBLIC_URL = 'http://dev.sso.lizard.net/'
-SSO_SERVER_PUBLIC_URL_V2 = 'http://dev.sso.lizard.net/api/v2/'
+SSO_SERVER_API_START_URL = 'http://dev.sso.lizard.net/api/v2/'
 # URL used for server-to-server communication
 # Note: needs a trailing slash
 SSO_SERVER_PRIVATE_URL = 'http://dev.sso.lizard.net:9874/'