+++ lizard_auth_client/views.py
@@ -139,7 +139,7 @@ class JWTLoginView(View):
             })
 
         # Build an absolute URL pointing to the SSO server out of it.
-        url = urljoin(settings.SSO_SERVER_PUBLIC_URL_V2, 'authenticate/')
+        url = urljoin(settings.SSO_SERVER_API_START_URL, 'authenticate/')
         url_with_params = '%s?%s' % (url, query_string)
         return HttpResponseRedirect(url_with_params)
 
@@ -242,7 +242,7 @@ class JWTLogoutView(View):
             'key': settings.SSO_KEY
             })
 
-        url = urljoin(settings.SSO_SERVER_PUBLIC_URL_V2, 'logout/')
+        url = urljoin(settings.SSO_SERVER_API_START_URL, 'logout/')
         url = '%s?%s' % (url, query_string)
 
         # send the redirect response